\name{PPCKO-package}
\alias{PPCKO-package}
\alias{PPCKO}
\docType{package}
\title{
\packageTitle{PPCKO}
}
\description{
\packageDescription{PPCKO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PPCKO}
\packageIndices{PPCKO}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PPCKO}

Maintainer: \packageMaintainer{PPCKO}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
